svg/coordinate-systems/parsing/vector-effect-computed.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Vector effects: getComputedStyle().vectorEffect</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/coords.html#VectorEffectProperty"/>
    <h:meta name="assert" content="vector-effect computed value is as specified."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/computed-testcommon.js"/>
  <script><![CDATA[

test_computed_value("vector-effect", "none");
test_computed_value("vector-effect", "non-scaling-stroke");

// These values are at risk: https://github.com/w3c/svgwg/issues/306
test_computed_value("vector-effect", "non-scaling-size");
test_computed_value("vector-effect", "non-rotation");
test_computed_value("vector-effect", "fixed-position");

  ]]></script>
</svg>
